menambah fungsi checkuser pada controller user

+++ Backend/Controllers/userController.js
@@ -7,6 +7,11 @@ async function addUser(name,password,interest){
         interest:interest,
         blogs:[]
     })
+    const userAvailability = await checkUser(name,password);
+    if(userAvailability==true){
+        console.log("User sudah ada");
+        return;
+    }
     await newUser.save();
     return;
 }
@@ -15,6 +20,11 @@ async function deleteUser(name,password){
     const identity = {
         userName:name,
         password:password
+    }
+    const userAvailability = await checkUser(name,password);
+    if(userAvailability==false){
+        console.log(`Tidak ada user dengan nama ${name}`);
+        return;
     }
     await user.deleteOne(identity);
     return;
@@ -33,5 +43,23 @@ async function fetchAll(){
     const allUser = await user.find();
     return allUser;
 }
+
+async function checkUser(name,password){
+    const identity={
+        userName:name,
+        password:password
+    }
+    const status = await user.findOne(identity);
+    if(status!=null){
+        return true;
+    }else{
+        return false;
+    }
+}
+
+export default{
+    addUser,deleteUser,fetchUser,fetchAll,checkUser
+}
+
